delete check write permissions

+++ packages/server/routes/delete.js
@@ -11,13 +11,14 @@ module.exports = router;
 router.post(
   "/:name/:id",
   setTenant,
-  loggedIn,
   error_catcher(async (req, res) => {
     const { name, id } = req.params;
     const { redirect } = req.query;
     const table = await Table.findOne({ name });
-    await table.deleteRows({ id });
+    const role = req.isAuthenticated() ? req.user.role_id : 10;
 
+    if (role <= table.min_role_write) await table.deleteRows({ id });
+    else req.flash("error", `Not allowed to write to table ${table.name}`);
     res.redirect(redirect || `/list/${table.name}`);
   })
 );
